+++ Entitas.Unity.VisualDebugging/Assets/Tests/Systems/SlowSystem.cs
@@ -3,7 +3,7 @@ using System.Threading;
 
 public class SlowSystem : IExecuteSystem {
     public void Execute() {
-        Thread.Sleep(12);
+        Thread.Sleep(4);
     }
 }