+++ Tests/Tests/Entitas.CodeGenerator/Fixtures/TestReactiveSystem.cs
@@ -3,7 +3,7 @@ using Entitas;
 
 public class TestReactiveSystem : IReactiveSystem {
 
-    public TriggerOnEvent trigger { get { return default(TriggerOnEvent); } }
+    public TriggerOnEvent trigger { get { return Matcher.AllOf(0).OnEntityAdded(); } }
 
     public void Execute(List<Entity> entities) {
     }