+++ build.sh
@@ -5,6 +5,8 @@ source build_commands.sh
 
 collect_sources
 update_project_dependencies
+generateProjectFiles
 build
 runTests
+collect_misc_files
 create_zip